+++ CHANGELOG.md
@@ -12,6 +12,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Comprehensive tests for context matching.
+
+### Fixed
+
+- Moved buffer append for before-context to after all match processing logic to ensure the buffer only contains the lines before the match.
+
+## [0.3.8] - 2025-10-25
+
+### Added
+
 - Adding `after_context` and `before_context` params to `search_multiple_files()` function and passing them through calls to `search_file()`.
 - Adding `after_context` and `before_context` params to `search_directory_recursively()` function and passing them through calls to `search_file()`.